test: split project onboarding browser journey
